Only push symbol packages for prereleases
Symbol packages include the .dll alongside the .pdb, so if you don't
need the .pdb then the only downside is a larger-than-needed download.
But since prerelease packages are what you're most likely to be trying
to debug, it's helpful to have the .pdb included.

This also avoids the "Package 'Foo.0.2.1.symbols.nupkg' already
exists" error when the "Foo.0.2.1.nupkg" package was uploaded earlier in
the process.