Make IChangeContext interface (#26)
* Create IChangeContext interface in core

Commit property in interface is CommitBase rather than Commit.

* Include harmony core in global usings
* Remove now-unneeded `using SIL.Harmony.Core` lines
* Create IObjectSnapshot interface in core

ObjectSnapshot relies on too many things from the DB to be moved to core
easily, so we leave it in SIL.Harmony.Db and create an interface in core
instead. Then the only difference is that the type of Commit will be
different depending on whether you access it via the interface (in which
case you'll get a CommitBase from core) or via the implementation
class (in which case you'll get a Commit object from SIL.Harmony).

* Tweak IChangeContext to use IObjectSnapshot
* Use interface in NewEntity and ApplyChange

The Change.NewEntity and Change.ApplyChange methods, as well as the
overriding/implementing methods in all Change subclasses, should now
take an IChangeContext so that it can be mocked for testing.

* Update RemoveReference methods to use CommitBase

Now that IObjectBase uses CommitBase in RemoveReference, its
implementing classes need to use the base type too.

* Pull GetObjectsReferencing into IChangeContext 
* Remove un-needed GetCurrent method from ChangeContext

---------

Co-authored-by: Kevin Hahn <kevin_hahn@sil.org>